requirements.txt
torch
torchvision
ultralytics
numpy
pillow
pycocotools
diffusers